chore: separate function process code
